\name{U2OS}
\docType{data}
\alias{U2OS}
\title{Luminescence measurements from U2-OS osteosarcoma cells}
\description{
This data set contains luminescence traces from osteosarcoma cells stably
expressing a Bmal1-luciferase reporter. In each experiment the cells were
transfected with siRNAs against circadian clock components.
}
\usage{U2OS}
\format{
A list containing the following time series objects: Fig.S1, Fig.S2AB,
Fig.S2C, Fig.S2D, Fig.S2E, Fig.S2F, Fig.S2GI, Fig.S2H, siBmal1, siBmal1\_Cry1,
siClock, siCry1\_Cry2, siCry1Cry2. Each column of these objects contains 721
measurements spaced 10 minutes apart. A very small proportion of missing data
points (less than 1 per 4,000) have been imputed by linear interpolation.
}
\references{
Baggs, J. E., Price, T. S., DeHaro L., Panda, S., FitzGerald, G. A., and Hogenesch, J. B.
\emph{Elucidating network structures of the circadian clock underlying robustness.}
Manuscript submitted for publication.
}
\keyword{datasets}
